requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-wtf==0.15.1
mysql-connector-python==8.0.26
selenium==4.1.0
python-dotenv==0.19.2
webdriver-manager==3.5.3